% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.gps.R
\name{best.sprint}
\alias{best.sprint}
\alias{gps.best.sprint}
\title{Best Sprint}
\usage{
best.sprint(
  game.speed,
  min.speed = 0.3,
  max.speed.threshold = 95,
  sample.rate = 10
)

gps.best.sprint(
  game.speed,
  min.speed = 0.3,
  max.speed.threshold = 95,
  sample.rate = 10
)
}
\arguments{
\item{game.speed}{The player's tracking data speed vector}

\item{min.speed}{The minimum speed that the function should consider the start of the sprint}

\item{max.speed.threshold}{The percent of the player's top speed reached to be considered the end of the sprint}

\item{sample.rate}{The game data sample rate (Hz)}
}
\value{
A data.frame with the player's best-sprint observations
}
\description{
This function returns the player's best-sprint from the data set.
}
\details{
This function identifies a player's best-sprint which is defined as the sprint that took the least amount
of time to reach top speed. The caveat being that this sprint does not need to begin at a dead-start.

The user sets the minimum speed that the function should consider the start of the sprint. Since players do
not often reach max speed during games and practices, the \code{max.speed.threshold} argument allows practitioners
to set the percent of top speed that the function should consider as the end of the sprint. Through trial-and-error,
the minimum speed of 0.3 m/s and 95% of top speed are appropriate values to build a player's in-game sprint
ability.
}
\section{Functions}{
\itemize{
\item \code{gps.best.sprint}: An alternate function call specifically for player tracking data
}}

