% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation.fvp}
\alias{citation.fvp}
\alias{cite.fvp}
\alias{reference.fvp}
\title{Cite {fvp}}
\usage{
citation.fvp()

cite.fvp()

reference.fvp()
}
\value{

}
\description{
This function returns a formatted citation for the {fvp} package
}
\section{Functions}{
\itemize{
\item \code{cite.fvp}: An alternative function call

\item \code{reference.fvp}: An alternative function call
}}

