% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.players.R
\name{compare.player.profiles}
\alias{compare.player.profiles}
\alias{compare.player.splits}
\title{Compare Players

This function returns a data.frame that compares players' mechanical sprint abilities using either their
player profiles.}
\usage{
compare.player.profiles(player.profiles = list())

compare.player.splits(player.profiles = list(), distance = c(9.14, 18.3, 36.6))
}
\arguments{
\item{player.profiles}{The players' profiles as a list}
}
\value{
A comparison of players' profiles or speed, acceleration, and split times at different distances
}
\description{
Compare Players

This function returns a data.frame that compares players' mechanical sprint abilities using either their
player profiles.
}
\section{Functions}{
\itemize{
\item \code{compare.player.splits}: Compare player split times at given distances
}}

