% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.to.R
\name{convert.to.metric}
\alias{convert.to.metric}
\alias{convert.to.meters}
\alias{convert.to.seconds}
\title{Convert Metric}
\usage{
convert.to.metric(value, units = "m/s")

convert.to.meters(value, units = c("ft", "km", "m", "mi", "yd"))

convert.to.seconds(units = c("second, minute, hour"))
}
\arguments{
\item{value}{A number}

\item{units}{The units corresponding to the number}
}
\value{

}
\description{
The functions in this package are optimized when values are in metric. This function provides an easy way
of converting values to m/s or m/s/s.
}
\section{Functions}{
\itemize{
\item \code{convert.to.meters}: Conversion to meters

\item \code{convert.to.seconds}: Conversion to seconds
}}

