% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.player.R
\name{data.player}
\alias{data.player}
\alias{fv.data.player}
\alias{fvp.data.player}
\title{Player Data}
\usage{
data.player(
  player.height,
  body.mass,
  lower.limb.length,
  initial.height,
  push.off.angle = 90,
  ambient.temp = 25,
  air.pressure = 760
)

fv.data.player(
  body.mass,
  lower.limb.length,
  initial.height,
  push.off.angle = 90
)

fvp.data.player(
  player.height,
  body.mass,
  ambient.temp = 25,
  air.pressure = 760
)
}
\arguments{
\item{player.height}{The player's height (meters)}

\item{body.mass}{The player's mass (kg)}

\item{lower.limb.length}{The player's lower-limb length (meters)}

\item{initial.height}{The player's height (meters)}

\item{push.off.angle}{The angle at which the player pushes off (degrees)}

\item{ambient.temp}{The air temperature at the time of testing (Celsius)}

\item{air.pressure}{The air pressure at the time of testing (mmHg)}
}
\value{
A data.frame that contains a player's anthropometric data and testing day weather
}
\description{
This is a generic function that supports a player's anthropometric data and testing day weather. The data
in this data set can be subsequently utilized in the functions \code{fv.results.model} and \code{fvp.results.model}.
}
\section{Functions}{
\itemize{
\item \code{fv.data.player}: A data.frame required for jump-based force-velocity analyses

\item \code{fvp.data.player}: Returns a data.frame required for sprint-based force-velocity-power analyses
}}

\note{
To properly implement jump-based force-velocity and sprint-based force-velocity-power profiling models,
measurements must be in metric.
}
