% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.testing.R
\name{data.testing}
\alias{data.testing}
\alias{fv.data.testing}
\title{Player Testing Data}
\usage{
data.testing(
  condition = percent.bw,
  additional.mass = add.mass,
  trial.one,
  trial.two,
  trial.three = NA
)

fv.data.testing(
  player.data,
  percent.bw,
  add.mass,
  trial.one,
  trial.two,
  trial.three = NA
)
}
\arguments{
\item{condition}{The percentage of body weight}

\item{additional.mass}{Mass added to the player}

\item{trial.one}{A vector containing results for trial 1 of all conditions}

\item{trial.two}{A vector containing results for trial 2 of all conditions}

\item{trial.three}{A vector containing results for trial 3 of all conditions}
}
\value{
A data.frame containing pertinent information for further force-velocity analyses
}
\description{
This is a generic function that builds a data.frame containing the data necessary for subsequent force-velocity
analyses. This function is to be used in conjunction with \code{_.data.testing()} when returning force-velocity
player profiles and force-velocity model results.
}
\section{Functions}{
\itemize{
\item \code{fv.data.testing}: A data.frame for subsequent jump-based force-velocity models
}}

\note{
The model requires at least two trials. A third trial is encouraged and should be input for the argument
\code{trial.three = NA}.

To properly implement force-velocity models, measurements must be in metric.
}
