% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.player.R
\name{fv.player.profile}
\alias{fv.player.profile}
\title{Force-Velocity Player Profile}
\usage{
fv.player.profile(player.name = NA, player.data, testing.data)
}
\arguments{
\item{player.name}{The player's name}

\item{player.data}{The player's anthopometric data}

\item{testing.data}{The athlete's jump-based testing data}
}
\value{
A player's jump-based force-velocity profile
}
\description{
This function return's a player's force-velocity profile. The values returned can be subsequently used to
build force-velocity plots and compare players' abilities.#'
}
