% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvp.fns.R
\name{fvp.fns}
\alias{fvp.fns}
\alias{air.force}
\alias{air.friction}
\alias{air.friction.coef}
\alias{air.rho}
\alias{force.horizontal}
\alias{power.horizontal}
\alias{force.resultant}
\alias{ratio.of.force}
\title{Force-Velocity-Power Functions}
\usage{
fvp.fns()

air.force(air.friction.coefficient, player.speed)

air.friction(player.height, player.mass)

air.friction.coef(player.data)

air.rho(ambient.temp, air.pressure)

force.horizontal(player.data, player.accel)

power.horizontal(player.speed, horiz.net.force)

force.resultant(horiz.force, body.mass)

ratio.of.force(horiz.force, result.force)
}
\value{
Values used when modeling a player's force-velocity-power profile
}
\description{
These are helper functions that are used when modeling a player's force-velocity-power profile.
}
\section{Functions}{
\itemize{
\item \code{air.force}: Force of air against the player

\item \code{air.friction}: Friction from air applied to the player

\item \code{air.friction.coef}: Air friction coefficient

\item \code{air.rho}: Air functions RHO constant

\item \code{force.horizontal}: Horizontal force applied by the athlete

\item \code{power.horizontal}: Horizontal power produced by the athlete

\item \code{force.resultant}: Total force vector

\item \code{ratio.of.force}: Ratio of horizontal to total force
}}

