% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvp.player.R
\name{fvp.player.profile}
\alias{fvp.player.profile}
\title{FVP Player Profile}
\usage{
fvp.player.profile(player.name = NA, player.data, player.profile)
}
\arguments{
\item{player.name}{The player's name}

\item{player.data}{The player's anthropometric data and testing day weather}

\item{player.profile}{The player's previously created profile}
}
\value{

}
\description{
This function returns a player's force-velocity-power profile
}
\details{
This function returns an athlete's complete force-velocity-power profile which includes:
- The player's modeled max force and max normalized force
- The player's modeled slope
- The player's modeled max power and max normalized power
- Ratio of forces
- The player's modeled max velocity
}
\note{
The \code{player.profile} argument takes on player profiles created using the \code{_.player.profile()} functions
}
