% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.gps.R
\name{gps.nls}
\alias{gps.nls}
\title{Fit Best-Sprint}
\usage{
gps.nls(sprint.data)
}
\arguments{
\item{sprint.data}{The player's best sprint data that is returned from the \code{best.sprint()} function}
}
\value{
An NLS with an optimized acceleration constant (player tau)
}
\description{
This function uses an optimization function to fit the athlete's best-sprint data. The function utilizes
the same function as that found in \code{speed.time()} and optimizes the player's acceleration constant to
minimize MSE.
}
