% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.splits.R
\name{player.splits}
\alias{player.splits}
\alias{gps.player.splits}
\alias{sa.player.splits}
\alias{scout.player.splits}
\title{Player Split Times}
\usage{
player.splits(player.profile, distance = c(9.14, 18.3, 36.6))

gps.player.splits(player.profile, distance = c(9.14, 18.3, 36.6))

sa.player.splits(player.profile, distance = c(9.14, 18.3, 36.6))

scout.player.splits(player.profile, distance = c(9.14, 18.3, 36.6))
}
\arguments{
\item{player.profile}{A player's sprint profile}

\item{distance}{The distance that corresponds with the split times recorded}
}
\value{

}
\description{
This function returns projected speed, acceleration, and split times based on their mechanical sprint abilities.
}
\section{Functions}{
\itemize{
\item \code{gps.player.splits}: Modeled speed, acceleration, and time observations from player tracking data

\item \code{sa.player.splits}: Modeled speed, acceleration, and time observations from speed-acceleration data

\item \code{scout.player.splits}: Speed, acceleration, and time observations from modeled sprint data
}}

