% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fitted.R
\name{results.fitted}
\alias{results.fitted}
\alias{sa.results.fitted}
\title{Fitted Results}
\usage{
results.fitted(game.data, r2 = 0.95)

sa.results.fitted(game.data, r2 = 0.95)
}
\arguments{
\item{game.data}{A player's game data}

\item{r2}{The minimum r^2 when fitting the model}
}
\value{
A data frame that returns fitted speed-acceleration observations
}
\description{
This function returns the fitted speed-acceleration results. This data can then be used to build an athlete's speed-acceleration
linear model.
}
\details{
This function is complimented by \code{sa.results.observed()}.

The observations provided by \code{sa.results.observed()} are fit to a linear model. A looped function removes the
observation with the largest residual, re-fitting the linear model, and assessing whether the model has reached an
r^2 of equal to, or greater than, that provided by the user. The r^2 value provided (0.95) corresponds with what
is suggested in the literature.
}
\section{Functions}{
\itemize{
\item \code{sa.results.fitted}: Fitted speed-acceleration observations
}}

