% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.game.R
\name{results.game}
\alias{results.game}
\alias{gps.results.game}
\title{Optimized Best Sprint Results}
\usage{
results.game(game.data, sample.rate = 10)

gps.results.game(game.data, sample.rate = 10)
}
\arguments{
\item{game.data}{The player's cleaned game data}

\item{sample.rate}{The game data sample rate (Hz)}
}
\value{
A data.frame that fits an optimized speed-time model to the athlete's best sprint
}
\description{
This function compliments \code{gps.results.observed()} in that it fits an optimized speed-time model to a
player's observed data. The data.frame returned allows practitioners to build plots displaying an athlete's in-game
sprint abilities.
}
\section{Functions}{
\itemize{
\item \code{gps.results.game}: A player's in-game best-sprint modeled data
}}

