% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.model.R
\name{results.model}
\alias{results.model}
\alias{fv.results.model}
\alias{gps.results.model}
\alias{sa.results.model}
\alias{scout.results.model}
\alias{fvp.results.model}
\title{Model Results}
\usage{
results.model(
  player.data,
  testing.data,
  game.data,
  push.off.angle = 90,
  sample.rate = 10,
  r2 = 0.95
)

fv.results.model(player.data, testing.data, push.off.angle = 90)

gps.results.model(game.data, sample.rate = 10)

sa.results.model(game.data, r2 = 0.95)

scout.results.model(sprint.data)

fvp.results.model(player.data, player.profile, sprint.duration = 5)
}
\arguments{
\item{player.data}{The player's anthropometric data and testing day weather}

\item{testing.data}{The player's force-velocity testing data}

\item{game.data}{The player's cleaned game data}

\item{push.off.angle}{Push-off angle for force-velocity testing (degrees)}

\item{sample.rate}{The game data sample rate (Hz)}

\item{r2}{The minimum r^2 when fitting the model}
}
\value{

}
\description{
\code{_.results.model()} functions return a data.frame that contains all observations that are required to build
plots from the resulting models. Unlike \code{_.player.profile()}, a player's abilities are not returned concisely.
Rather, this function was built to provide practitioners the ability to plot player's modeled abilities for
simple visualization.
}
\details{
\code{_.results._()} functions are built with consistency in mind. Therefore, the majority of them take on a player's
observed or tested data. The exception to this rule is \code{fvp._()} functions because secondary analyses are
required before returning a player's profile and modeled data.

The \code{_.results.game()}, \code{_.results.fitted()}, and \code{_.results.observed()} return similar data.frames but
behave differently on the back-end. Therefore, it is suggested that these functions be used for their intended purposes.
}
\section{Functions}{
\itemize{
\item \code{fv.results.model}: Force-velocity modeled observations

\item \code{gps.results.model}: Player tracking data modeled observations

\item \code{sa.results.model}: Speed-acceleration modeled observations

\item \code{scout.results.model}: Sprint modeled observations

\item \code{fvp.results.model}: Force-velocity-power modeled observations
}}

