% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.sa.R
\name{sa.lm.observed}
\alias{sa.lm.observed}
\alias{sa.lm.fitted}
\title{Speed-Accel Linear Model Observations}
\usage{
sa.lm.observed(game.data)

sa.lm.fitted(game.data, r2 = 0.95)
}
\arguments{
\item{game.data}{The player's cleaned game data}
}
\value{
A data.frame with observations that meet speed-acceleration inclusion criteria
}
\description{
This function returns the observations that fit inclusion criteria when building a player's
speed-acceleration profile and the observations kept that subsequently build the linear model.
}
\section{Functions}{
\itemize{
\item \code{sa.lm.fitted}: A data.frame with observations that were maintained to build the speed-acceleration linear model
}}

