% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.as.R
\name{save.as.csv}
\alias{save.as.csv}
\alias{save.as.excel}
\alias{save.as.rdata}
\alias{save.as.rds}
\title{Save As}
\usage{
save.as.csv(data.set, file.name)

save.as.excel(data.set, file.name)

save.as.rdata(data.set, file.name)

save.as.rds(data.set, file.name)
}
\arguments{
\item{data.set}{The player's returned data.frame}

\item{file.name}{The intended file name}
}
\description{
These functions allow users to save results returned from \code{_.player.profile()}, \code{_.results._()}, and
\code{compare.player._()} functions as a Microsoft Excel document, CSV document, or RDS (Rdata) file
}
\section{Functions}{
\itemize{
\item \code{save.as.excel}: Saves a player's data as a .csv file

\item \code{save.as.rdata}: Saves a player's data as an Rdata file

\item \code{save.as.rds}: Saves a player's data as an RDS file
}}

