% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tracking.R
\name{scout.data}
\alias{scout.data}
\title{Split Time Data}
\usage{
scout.data(distance = c(9.14, 18.3, 36.6), split.time)
}
\arguments{
\item{distance}{Distance traveled}

\item{split.time}{Time to cover the given distance}
}
\value{
The distance and time data.frame required for future analyses
}
\description{
This function builds the data.frame necessary to fit an athlete's mechanical sprint ability optimization
function. The argument \code{distance = c(9.14, 18.3, 36.6)} contains distances that are equivalent to the
10, 20, and 40 yard splits.
}
\details{
Although this function can handle fewer split times, it is not recommended to do so. Subsequent analyses
rely on models that were built, and have been validated, utilizing the distances provided.
}
\note{
The input game data is assumed to be in metric. Speed must be in m/s and acceleration in m/s/s. If it is not,
values that are returned are not representative of a player's true sprint potential. If your data needs to be converted,
use the \code{convert.to.metric} function.
}
