% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint.fns.R
\name{sprint.fns}
\alias{sprint.fns}
\alias{accel.time}
\alias{distance.speed}
\alias{distance.time}
\alias{midsprint}
\alias{speed.time}
\alias{time.distance}
\alias{time.speed}
\alias{time.to.dist.given.speed}
\alias{time.to.position}
\title{Mechanical Sprint Abilities}
\usage{
sprint.fns()

accel.time(player.profile, split.time)

distance.speed(player.profile, current.speed)

distance.time(player.profile, split.time)

midsprint(player.profile, current.speed, distance)

speed.time(player.profile, split.time)

\method{time}{distance}(player.profile, distance)

\method{time}{speed}(player.profile, current.speed)

\method{time}{to.dist.given.speed}(player.profile, current.speed, distance)

\method{time}{to.position}(player.profile, current.speed, distance)
}
\value{
Speed, acceleration, distance, or time values corresponding to an athlete's sprint abilities
}
\description{
A set of functions that return speed, acceleration, distance, and time values as functions of each other. These are
helper function when building \code{_.results._()} data.frames.
}
\details{
Thank you to Dani Chu for his contributions to many of these functions.
}
